requirements.txt
# OCR and Image Processing
pytesseract>=0.3.10
Pillow>=10.1.0
pdf2image>=1.16.3
pymupdf>=1.24.0

# Word Document Processing
python-docx>=1.1.0

# AI Model
openai>=1.3.7

# Data Processing
pandas>=2.1.4

# Config
python-dotenv>=1.0.0

# Progress
tqdm>=4.66.1

# Logging
colorlog>=6.8.0

# Web
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.1

# PDF Generation
reportlab>=4.0.0
